+++ src/ws/BybitPublicStream.types.ts
@@ -1,3 +1,5 @@
+import type { ReliableWebSocket } from '@solncebro/websocket-engine';
+
 import type { ExchangeLogger } from '../types/common';
 import type { BybitBaseWebSocketMessage } from './bybitWebSocketUtils.types';
 
@@ -11,4 +13,12 @@ export interface BybitPublicStreamArgs {
 export interface BybitWebSocketMessage extends BybitBaseWebSocketMessage {
   topic?: string;
   ret_code?: number;
+}
+
+export interface BybitConnection {
+  webSocket: ReliableWebSocket<BybitWebSocketMessage>;
+  label: string;
+  topicList: string[];
+  dynamicTopicList: string[];
+  url: string;
 }